requirements.txt
python-dotenv
python-dateutil

openmeteo-requests
requests-cache
retry-requests

pandas
scikit-learn
xgboost
torch
torchvision

fastapi[standard]

black
